requirements/requirements.txt
numpy
pandas
openpyxl
scikit-learn
scipy
umap-learn
xgboost
catboost
lightgbm

torch
tensorflow

matplotlib
seaborn
plotly

shap
optuna
wandb

jupyter
ipykernel

